Processing requests with the site
In the http body, the arguments have two values request (accepts) and response (sends).
Inside the arrow function, a condition is formed: if the value of request.method will work over the GET protocol, then the application will return data to the user, and if POST, the application will accept the data sent by the user.