Collecting data from the buffer
In the case of POST, the response object will also execute the writeHead, write and end methods. However, data will be received via request. The page and the server exchange data of String type, and are transmitted as a buffer.
To contain such data, an array is required that can receive data or chunks element by element. The request.on method listens for the 'data' buffer. Using the body.push method, the array adds new data from data element by element to the end using the Buffer.from method.