Forming a response
The request.on method with the end argument indicates that all data has been received in its entirety. It remains only to decode them using the split method, which will split the data items by indices, where it will return the value at the second index. The received message is passed to massage.