Html page addressing
Created index.html about.html pages with links.
When building the page, a division by pages was formed if the address came '/' of the main page, and if '/about.html' of the additional page.
The readFile method that takes as the first argument
* path of the file from which to read data
path.join (__ dirname, 'views', 'about.html')
* the second encoding is 'utf-8'
* the third function, the arguments of which will be err read error and the variable content, which will contain the file data.
response.end (content)