Working with the console
When you run the program using the node console.js command, you can pass additional variables. This is how node console.js variable = value. They can be read through the global process.argv object and output through console.log (). The result will be the address of the running program and executable file. The passed variables are the next arguments.
A function is written that excludes the first two arguments and converts the string data of variables into array objects.