Express module routing
Formed two files index.html and about.html.
The express module allows you to form html pages. To do this, it has a get method, which as the first argument took '/ address' the address to which the page is accessed, and a method with arguments (req, res) that accepts and sends arguments. Inside the method, the about.html file from the views folder is sent to the res object through the sendFile method.