Handlebar engine connection
First, the handlebar module needs to be installed npm install express-handlebars. Then connect const exps = require ('express-handlebars') in the startup file. Next, you can create an hbs object of the exps class with configuration, where defaultLayout is the default layout name, extname specifies the abbreviated name of the engine.
To register the engine for the app server, you need to refer to the engine method, where the first argument is the name of the hbs engine, and the second is the hbs.engine value. Further, to use the hbs engine, the set method includes the name of the engine and the name of the folder where the html page will be located.
This completes the handlebar settings. After that, .html files must be renamed to .hbs. As a result, to display the page, the res object needs to use the render method, which will take the file name as an argument.