Saving new products to a file
A new product.js file has been implemented. Inside it, a Product class is created with internal fields:
             + title
             + price
             + img
             + id
and methods:
             + save()
             + getAll()
             + toJSON()
In the pages.js file, in the POST method of the routerPersonalArea object, a product instance of the Product class is instantiated with the parameters of the class instance, stern id. This class variable gets a random result from the uuid library.
The getAll() method reads the contents of the product.json file through the readFile method of the fs object.
The save() method first calls the getAll() method to retrieve the information already written. Further in the method, new objects are added to the created object through the push method, the argument of which is the toJSON() method.
The toJSON() method returns recently written local objects of the Product class.
The save() method ends with the fs object's writeFile method, which writes out the updated product object.
based on the result of the POST method, the page is redirected to the address /products.