Output of stored products
In the get method of the routerProducts object, the arrow method is made asynchronous.
Added the getAll () method of the Product class, the result of which is the product object. This object is added to the site arguments products.
A boolean expression is added to the products page in case there are no array instances.
If there are array objects, the #each construction is used internally, enumerating the products array.
Inside html markup of a block with a picture, name, price and a link to detailed information of the product is created.