Product view page
Added get method for product page.
The method uses a dynamic product id. The req.params method is used to get the id value.
To get one instance of the list, the getByID () method is used, the argument of which is the product id. The Product class method uses the getAll () method to load the entire list of products. After loading from this list, the required instance is found by id using the find () function.
After receiving all the data for the desired instance, it is sent to the 'product' page using the render method of the res object.
A new page product.hbs is created with html markup and css styles.