Layout page opening
When opening a link by id, a layout field has been added to the routerProducts object, which redirects the request to a new page located in the empty.hbs file.
In the products.hbs file, in the 'View product' link, the target = "_blank" property has been added, which opens the link in the browser's link.