Editing saved data
The data editing productEdit.hbs file was created based on the personalArea.hbs file. Additionally, existing values ​​have been added to the fields. We also added an <input> field by assigning an existing id value.
In order to get to the editing page, a second link to the editing page with the address "/products/{{id}}/edit" was added in the html markup of the products.hbs file.
In the pages.js routing file for the product page, a GET method has been added to send to the edit page, and a POST method to receive new edited object values.
The GET method checks the received req.query.allow value from the link. If it is negative, then the program is redirected to the home page.
Further, the product data is loaded using the getByID() method of the Product class, into which the req.params.id of the product is passed.
In the POST method, the update() method is executed, the argument of which is supplied with the new product data. Then there is a redirect to the catalog page.
The update() method works in the product.js file of the Product class as follows. First, all data is loaded through the getAll() method. After that, the required instance is found from the resulting list by id using the findIndex() method. According to the received id, new data is substituted into the list to replace the old ones. After the new list is received, it is written to the product.json file.