Removing products from the basket
To remove the product, the AJAX technology of asynchronous processing of requests to the site was used.
First, the html object '#pay' was received using the querySelector method of the global document class into the $pay query variable.
The request instance calls the addEventListener method, which listens for the 'click' action of clicking on the screen.
Inside the method, the condition is met. If the link event.target.classList.contains checks if there is such an object or, in other words, if a click was made on the 'js-remove' object, then the object identifier is copied by event.target.dataset.id to the id variable.
To process the AJAX request by the server, the fetch() method is used, which as the first argument takes the path along which the processing will be performed, the second the 'delete' method of processing.
Further, the fetch() method returns a Promise, therefore, you can use .then, in which we decode the response using the json() method.
The next use of .then will already receive a pay object, which will contain all the objects that need to be displayed in the cart table.
Using the map() method, the html markup with substituted product values ​​is formed element by element.
The result of the map() function is an html array that needs to be converted to a string using the join('') function.
The html array is entered into the HTML 'tbody' tag using the querySelector method and specifying .innerHTML that this is HTML markup.
The same is done with the price value, which is entered in the HTML '.price' tag as textContent.

To handle an AJAX request for the fetch() method, it must be processed from the server side in the pages.js file. To do this, the router routerCard using the delete() method creates a pay object created in an asynchronous function from the remove() method of the Pay class with the product id argument. After that, the pay object is also sent asynchronously back to the page.

For all the logic to work, the remove() method is used, which works as follows in the pay.js file. First, the fetch() method is used to load the list of products in the cart. After that, the index of the desired product from the list is found using the findIndex() function.
Next comes the condition. If the quantity of the sought product is equal to one, then it is removed from the list. If the quantity is more than one, then the number of copies of the product is decremented.
Then the removed product is deducted from the total.
At the end, the resulting list is written to the pay.json file.