Launching an application with a connected database
Included the 'mongoose' library and the URL variable from the password.js file.
The launch of the application has been moved to the asynchronous start () function, where the entire launch is placed in a try catch (err) construction.
The connection to the database is made using the connect () method of the mongoose object, where the first argument is the URL to the database.
Next, the application is launched.