Added a button to remove a product
On the product edit page, productEdit.hbs file, added form markup html markup with post method with product delete button.
Added a post () method to the routerProducts object to handle the event. Inside the method, the deleteOne () function is used, which takes as an argument the id of the product that needs to be deleted.