The method for adding a product to the cart
+ Added addToPay method in user.js file.
First, the method loads the entire list of products in the user's cart. Also finds the index from the cart list of the newly added product. If there is no such index, then this product is added to the list. If such a product is already in the basket, then the quantity of the product in the list is decremented.
After adding a new product, the list is saved in the basket field, after which the user data is saved to the database.
- The pay.js file has been removed.
+ Since all methods of working with the file have been removed, the connection to the pay.js file has been removed in the pages.js file, and the Pay.add() and Product.getByID() methods have been removed. Instead, the Product.findById() method from the 'mongoose' package and the req.user.addToPay() method from the user.js file described earlier have been made.