Display the list of korizny in the table
The method of sending data to the page has been changed in the pages.js file. Removed Pay.fetch() from the get() method of the routerCard object.
User data is loaded instead. The populate() method identifies the fields to be assigned, and the execPopulate() method is used for data format convenience.
Through the learned object, user processes the items field, using the map() method, where all the productId fields are obtained using _doc, and the count field.
To get the sum of all products, the reduce() method is applied to the pay object, with the total, iteration variable, and an initial value. The method contains the sum of all products multiplied by their quantity.
The resulting objects are sent to the products: pay, priceProduct: price fields.